requirements.txt
# Django and REST framework
Django==4.2.0
django-cors-headers==4.3.0
djangorestframework==3.14.0
requests==2.31.0
google-generativeai==0.3.0
clarifai-grpc==9.10.0
Pillow==10.0.0
python-dotenv==1.0.0
django-debug-toolbar==4.2.0
django-allauth==0.57.0
python-magic==0.4.27
pytz==2023.3
python-decouple==3.8 